move lib/ to platform/ and build and link correct dependencies
