+++ src/main.c
@@ -1,12 +1,12 @@
-#include "lib/sh68f90a/clock.h"
-#include "lib/sh68f90a/ldo.h"
-#include "lib/sh68f90a/watchdog.h"
-#include "lib/sh68f90a/delay.h"
-#include "lib/sh68f90a/isp.h"
-#include "lib/sh68f90a/uart.h"
-#include "lib/sh68f90a/gpio.h"
-#include "lib/sh68f90a/pwm.h"
-#include "lib/sh68f90a/usb.h"
+#include "platform/sh68f90a/clock.h"
+#include "platform/sh68f90a/ldo.h"
+#include "platform/sh68f90a/watchdog.h"
+#include "platform/sh68f90a/delay.h"
+#include "platform/sh68f90a/isp.h"
+#include "platform/sh68f90a/uart.h"
+#include "platform/sh68f90a/gpio.h"
+#include "platform/sh68f90a/pwm.h"
+#include "platform/sh68f90a/usb.h"
 #include "smk/debug.h"
 #include "smk/matrix.h"
 #include "smk/utils.h"
