+++ src/platform/sh68f90a/clock.h
