+++ src/platform/sh68f90a/usbregs.h
