+++ src/smk/matrix.c
@@ -1,9 +1,9 @@
 #include "matrix.h"
 #include "report.h"
 #include "debug.h"
-#include "../lib/sh68f90a/sh68f90a.h"
-#include "../lib/sh68f90a/pwm.h"
-#include "../lib/sh68f90a/delay.h"
+#include "../platform/sh68f90a/sh68f90a.h"
+#include "../platform/sh68f90a/pwm.h"
+#include "../platform/sh68f90a/delay.h"
 #include "../user/layout.h"
 #include "../user/indicators.h"
 #include "../user/matrix.h"
