+++ src/user/indicators.c
@@ -1,4 +1,4 @@
-#include "../indicators.h"
+#include "indicators.h"
 
 bool indicators_update_step(keyboard_state_t *keyboard, uint8_t current_step)
 {
