+++ src/user/layout.c
@@ -1,4 +1,4 @@
-#include "../layout.h"
+#include "layout.h"
 
 bool process_record_user(uint16_t keycode, bool key_pressed)
 {
